[Config] Fixed config values

+++ src/boards/Parafoil/StateMachines/WingController/WingController.h
@@ -102,7 +102,7 @@ public:
      * @param algorithm The algorithm with
      * all already done (e.g. steps already registered)
      */
-    void addAlgorithm(uint8_t id);
+    void addAlgorithm(int id);
 
     /**
      * @brief Construct a new Wing Controller object
